Hide the members page if the feature flag is disabled
